Download the necessary data from a remote store in a local

Download the necessary data from a remote store in a local

download_from_remote(local, remote, config)

Arguments

local
local rOstluft store. Probably a rOstluft::storage_local_rds()

remote
remote rOstluft store. Probably a rOstluft::storage_s3_rds()

config
config in list form from yaml::read_yaml()

Value

resulting tibble from local$put()

Examples

config <- get_example_config()
remote <- rOstluft::store_aqmet()
local <- rOstluft::storage_local_rds(
  "rf_example",
  format = rOstluft::format_rolf(),
  read.only = FALSE
)

#> Local store rf_example initialized under 'C:\Users\tom\AppData\Local\rOstluft\rf_example'

download_from_remote(local, remote, config)

#> Client error: (404) Not Found

#> First put to storage. Save columns types to C:/Users/tom/AppData/Local/rOstluft/rf_example/columns.rds

#> # A tibble: 156 x 6
#>     year interval site                parameter unit      n
#>    <dbl> <fct>    <fct>               <fct>     <fct> <int>
#>  1  2015 h1       Zch_Schimmelstrasse CO        mg/m3  8534
#>  2  2015 h1       Zch_Schimmelstrasse Hr        %Hr    8583
#>  3  2015 h1       Zch_Schimmelstrasse NO2       µg/m3  8524
#>  4  2015 h1       Zch_Schimmelstrasse NO        µg/m3  8524
#>  5  2015 h1       Zch_Schimmelstrasse NOx       ppb    8524
#>  6  2015 h1       Zch_Schimmelstrasse O3        µg/m3  8424
#>  7  2015 h1       Zch_Schimmelstrasse PM10      µg/m3  8489
#>  8  2015 h1       Zch_Schimmelstrasse PN        1/cm3   182
#>  9  2015 h1       Zch_Schimmelstrasse RainDur   min    8597
#> 10  2015 h1       Zch_Schimmelstrasse SO2       µg/m3  8502
#> # ... with 146 more rows

# tidy up
local$destroy("DELETE")

#> Store rf_example destroyed