Get data from store and rearange to wide format

Get data from store and rearange to wide format

prepare_observations_rf(config, store)

Arguments

config
in list form from yaml::read_yaml()

store
containing all the data

Value

list with items observations and n_observations as summary

Examples

config <- get_example_config()
store <- rOstluft::store_aqmet()
prepare_observations_rf(config, store)

#> Client error: (404) Not Found

#> $observations
#> # A tibble: 46,591 x 10
#>    date                   Hr     p RainSum StrGlo     T value    WD WD_local
#>    <dttm>              <dbl> <dbl>   <dbl>  <dbl> <dbl> <dbl> <dbl>    <dbl>
#>  1 2015-01-01 00:00:00  95.6  986.       0   0    -4.33  64.7  53.4    337. 
#>  2 2015-01-01 01:00:00  95.4  986.       0   0    -4.55  82.4  55.5    339. 
#>  3 2015-01-01 02:00:00  94.5  986.       0   0    -4.3   63.1  67.8    358. 
#>  4 2015-01-01 03:00:00  94.8  986.       0   0    -4.13  76.6  65.2    339. 
#>  5 2015-01-01 04:00:00  96.2  986.       0   0    -4.57  52.5  42.4    335. 
#>  6 2015-01-01 05:00:00  96.1  986.       0   0    -4.05  38.6  69.8    339. 
#>  7 2015-01-01 06:00:00  97.9  987.       0   0    -4     35.8  59.3    348. 
#>  8 2015-01-01 07:00:00  98.9  987.       0   0    -4.72  34.0  54.8    344. 
#>  9 2015-01-01 08:00:00  99.5  987.       0   4.17 -4.75  35.0  52.9     16.7
#> 10 2015-01-01 09:00:00  99.3  988.       0  35.3  -4.83  37.8  49.7    353. 
#> # ... with 46,581 more rows, and 1 more variable: WVs <dbl>
#> 
#> $n_observations
#> # A tibble: 54 x 5
#>    variable site             parameter  year     n
#>    <chr>    <chr>            <chr>     <dbl> <int>
#>  1 Hr       Zürich/Fluntern  Hr         2015  8760
#>  2 Hr       Zürich/Fluntern  Hr         2016  8784
#>  3 Hr       Zürich/Fluntern  Hr         2017  8760
#>  4 Hr       Zürich/Fluntern  Hr         2018  8760
#>  5 Hr       Zürich/Fluntern  Hr         2019  8760
#>  6 Hr       Zürich/Fluntern  Hr         2020  2767
#>  7 p        ZH_Tiefenbrunnen p          2015  8321
#>  8 p        ZH_Tiefenbrunnen p          2016  8781
#>  9 p        ZH_Tiefenbrunnen p          2017  8756
#> 10 p        ZH_Tiefenbrunnen p          2018  8745
#> # ... with 44 more rows
#>